Move ProgressCircle to custom-annotation example
